refactor(rust): move semantic version to its own module
