+++ sdk/rust/src/network_version_info.rs
@@ -20,31 +20,10 @@
 
 use hedera_proto::services;
 
-use crate::FromProtobuf;
-
-/// Hedera follows [semantic versioning](https://semver.org) for both the HAPI protobufs and
-/// the Services software.
-#[derive(Debug, Clone, serde::Serialize)]
-pub struct SemanticVersion {
-    /// Increases with incompatible API changes
-    pub major: u32,
-
-    /// Increases with backwards-compatible new functionality
-    pub minor: u32,
-
-    /// Increases with backwards-compatible bug fixes]
-    pub patch: u32,
-
-    /// A pre-release version MAY be denoted by appending a hyphen and a series of dot separated identifiers (https://semver.org/#spec-item-9);
-    /// so given a semver 0.14.0-alpha.1+21AF26D3, this field would contain ‘alpha.1’
-    #[serde(skip_serializing_if = "String::is_empty")]
-    pub prerelease: String,
-    /// Build metadata MAY be denoted by appending a plus sign and a series of dot separated identifiers
-    /// immediately following the patch or pre-release version (https://semver.org/#spec-item-10);
-    /// so given a semver 0.14.0-alpha.1+21AF26D3, this field would contain ‘21AF26D3’
-    #[serde(skip_serializing_if = "String::is_empty")]
-    pub build: String,
-}
+use crate::{
+    FromProtobuf,
+    SemanticVersion,
+};
 
 /// Versions of Hedera Services, and the protobuf schema.
 #[derive(Debug, Clone, serde::Serialize)]
@@ -69,21 +48,6 @@ impl FromProtobuf<services::response::Response> for NetworkVersionInfo {
         Ok(Self {
             protobuf_version: SemanticVersion::from_protobuf(protobuf_version)?,
             services_version: SemanticVersion::from_protobuf(services_version)?,
-        })
-    }
-}
-
-impl FromProtobuf<services::SemanticVersion> for SemanticVersion {
-    fn from_protobuf(pb: services::SemanticVersion) -> crate::Result<Self>
-    where
-        Self: Sized,
-    {
-        Ok(Self {
-            major: pb.major as u32,
-            minor: pb.minor as u32,
-            patch: pb.patch as u32,
-            prerelease: pb.pre,
-            build: pb.build,
         })
     }
 }
